requirements.txt
discord[voice]
pyyaml
discord-py-interactions
pynacl
requests
wave